refactor(modelprovider): centralize automatic model preferences
